Better GUI (#1)
* Added image queues

* Added ShimmerQueue

ShimmerQueue wraps the functionality for queueing images for the
shimmer frontend. It serves up new images with the highest priority,
and loops over old images when no new images are available.

* Beginning of a more flexible UI

Sort of using the model view controller pattern. The View being HTML/CSS, controller being the ShimmerNav.js object and the model being Shimmer.js. This should make it easier for the appliaction to support a couple different use cases.

* Shape dialog with freezing target selector

* Extracting 'subclasses' for tools

* Saving targets into targets.json serverside. single image only

* Export target data to targets.json on exit

* Adding current changes to resolve conflicts with #2

* Fixed merge errors

When merging, the atexit import was lost.

Also, the merge seemed to introduce a new bug(though it may have been
there before) where an image would not show after having targets
submitted the first time because some data was overwritten in the
server's targets array when the data was POSTed. This fixes that bug
by only writing the new target info into the array.