Fixed issue when submitting target
Originally, we only error check the x axis to ensure that the
x coordiates are arranged smallest to largest so we always grab the top
left corner.
This was only being done for the x coordinate so if we grabbed from the
bottom corner, our y cordinate will be larger which python array slicing
does not enjoy.
Added additional check in frontend to fix this issue